* Merge branches/tvo3: rts
git-svn-id: https://spring.clan-sy.com/svn/spring/trunk@1889 37977431-3df6-0310-b722-df95706aa16b

+++ rts/Game/Game.h
@@ -83,7 +83,7 @@ public:
 
 	CglList* showList;
 	Uint64 lastframe;
-	double totalGameTime;			//time in seconds, stops at game end
+	float totalGameTime;			//time in seconds, stops at game end
 
 	float maxUserSpeed;
 	float minUserSpeed;
